services/producer/requirements.txt
pandas==2.2.2
kafka-python==2.0.2
python-dateutil==2.9.0
